* telnetd/utility.c: Include <time.h>. (expand_line): Cast LINE to (char *).
* ftpd/auth.c: Include <string.h>
(auth_user) <default>: Cast LEN to (size_t).

* ftp/cmdtab.c: Include <string.h>
* libinetutils/xalloc_die.c: Include <stdlib.h>
* talk/get_addrs.c: Include <stdlib.h>, and <string.h>.
* talk/get_names.c: Include <stdlib.h>, and <string.h>.
* talk/init_disp.c: Include <stdlib.h>.
* talk/invite.c: Include <string.h>
* talk/look_up.c: Include <string.h>
* talk/talk.c: Include <stdlib.h>
* telnet/network.c: Include <stdlib.h>
* telnet/ring.c: Include <string.h>.
* telnet/terminal.c: Include <stdlib.h>.
* telnet/utilities.c: Include <stdlib.h>.

* rexecd/rexecd.c: Renamed all occurences of ASIN to A_SIN.

* libtelnet/encrypt.c, ftp/cmds.c, libls/util.c,
libls/stat_flags.c, libls/print.c, libls/ls.c, libls/fts.c,
libinetutils/tftpsubs.c, libinetutils/des_rw.c, ftpd/popen.c,
ftpd/logwtmp.c, ftpd/ftpcmd.y, whois/net.c, telnetd/termstat.c,
telnetd/state.c, telnetd/slc.c, telnet/utilities.c, telnet/main.c,
libtelnet/misc-proto.h, libtelnet/key-proto.h,
libtelnet/forward.c, rcp/util.c, talk/invite.c, telnet/externs.h:
Remove useless casts to (void).

* libtelnet/encrypt.c, libtelnet/enc_des.c, libtelnet/enc-proto.h,
libtelnet/auth.h, telnetd/utility.c, telnetd/telnetd.h,
telnetd/telnetd.c, telnet/ring.h, telnet/krb4-proto.h,
telnet/commands.c, rlogind/rlogind.c, libtelnet/encrypt.h,
libtelnet/kerberos.c, libtelnet/read_passwd.c: Remove all
references to P().

+++ ftpd/auth.c
@@ -4,6 +4,7 @@
 
 #include <sys/types.h>
 #include <stdlib.h>
+#include <string.h>
 #include <unistd.h>
 #include <stdio.h>
 #include <pwd.h>
@@ -63,7 +64,7 @@ auth_user (const char *name, struct credentials *pcred)
 	size_t len;
 	if (pcred->message)
 	  free (pcred->message);
-	len = 64 + strlen (name);
+	len = (size_t) (64 + strlen (name));
 	pcred->message = malloc (len);
 	if (pcred->message == NULL)
 	  return -1;
